new changes

+++ src/main/java/com/all/in/one/allinOne/repository/ModelRepository.java
@@ -4,5 +4,5 @@ import com.all.in.one.allinOne.entity.City;
 import com.all.in.one.allinOne.entity.Model;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ModelRepository extends JpaRepository<Model, Integer> {
+public interface ModelRepository extends JpaRepository<Model, Long> {
 }
